+++ CHANGELOG.md
@@ -72,6 +72,7 @@ VS Code v1.56
 ### Development
 
 - chore: ignore updates to microsoft/playwright-github-action
+- fix(socket): use xdgBasedir.runtime instead of tmp #3304 @jsjoeio
 
 ## 3.10.0